+++ src/core/scores.cpp
@@ -225,7 +225,7 @@ errr top_twenty(player_type *current_player_ptr)
     sprintf(the_score.cur_lev, "%3d", MIN((uint16_t)current_player_ptr->lev, 999));
     sprintf(the_score.cur_dun, "%3d", (int)current_player_ptr->current_floor_ptr->dun_level);
     sprintf(the_score.max_lev, "%3d", MIN((uint16_t)current_player_ptr->max_plv, 999));
-    sprintf(the_score.max_dun, "%3d", (int)max_dlv[current_player_ptr->dungeon_idx]);
+    sprintf(the_score.max_dun, "%3d", (int)max_dlv[static_cast<int>(current_player_ptr->dungeon_idx)]);
 
     /* Save the cause of death (31 chars) */
     if (strlen(current_player_ptr->died_from) >= sizeof(the_score.how)) {
@@ -328,7 +328,7 @@ errr predict_score(player_type *current_player_ptr)
     sprintf(the_score.cur_lev, "%3d", MIN((uint16_t)current_player_ptr->lev, 999));
     sprintf(the_score.cur_dun, "%3d", (int)current_player_ptr->current_floor_ptr->dun_level);
     sprintf(the_score.max_lev, "%3d", MIN((uint16_t)current_player_ptr->max_plv, 999));
-    sprintf(the_score.max_dun, "%3d", (int)max_dlv[current_player_ptr->dungeon_idx]);
+    sprintf(the_score.max_dun, "%3d", (int)max_dlv[static_cast<int>(current_player_ptr->dungeon_idx)]);
 
     /* Hack -- no cause of death */
     /* まだ死んでいないときの識別文字 */
