+++ src/effect/effect-feature.cpp
@@ -318,7 +318,7 @@ bool affect_feature(player_type *caster_ptr, MONSTER_IDX who, POSITION r, POSITI
     }
     case GF_LITE_WEAK:
     case GF_LITE: {
-        if (d_info[caster_ptr->dungeon_idx].flags.has(DF::DARKNESS))
+        if (d_info[static_cast<int>(caster_ptr->dungeon_idx)].flags.has(DF::DARKNESS))
             break;
 
         g_ptr->info |= (CAVE_GLOW);
