+++ src/floor/cave-generator.cpp
@@ -191,7 +191,7 @@ static bool make_one_floor(player_type *player_ptr, dun_data_type *dd_ptr, dunge
 {
     floor_type *floor_ptr = player_ptr->current_floor_ptr;
 
-    if (d_info[floor_ptr->dungeon_idx].flags.has(DF::NO_ROOM)) {
+    if (d_info[static_cast<int>(floor_ptr->dungeon_idx)].flags.has(DF::NO_ROOM)) {
         make_only_tunnel_points(floor_ptr, dd_ptr);
     } else {
         if (!generate_rooms(player_ptr, dd_ptr)) {
@@ -394,7 +394,7 @@ bool cave_gen(player_type *player_ptr, concptr *why)
             dd_ptr->room_map[y][x] = false;
 
     dd_ptr->cent_n = 0;
-    dungeon_type *d_ptr = &d_info[floor_ptr->dungeon_idx];
+    dungeon_type *d_ptr = &d_info[static_cast<int>(floor_ptr->dungeon_idx)];
     if (ironman_empty_levels || (d_ptr->flags.has(DF::ARENA) && (empty_levels && one_in_(EMPTY_LEVEL)))) {
         dd_ptr->empty_level = true;
         msg_print_wizard(player_ptr, CHEAT_DUNGEON, _("アリーナレベルを生成。", "Arena level."));
