+++ src/io-dump/character-dump.cpp
@@ -160,20 +160,20 @@ static void dump_aux_last_message(player_type *creature_ptr, FILE *fff)
 static void dump_aux_recall(FILE *fff)
 {
     fprintf(fff, _("\n  [帰還場所]\n\n", "\n  [Recall Depth]\n\n"));
-    for (int y = 1; y < current_world_ptr->max_d_idx; y++) {
+    for (int y = 1; y < static_cast<int>(current_world_ptr->max_d_idx); y++) {
         bool seiha = false;
 
-        if (!d_info[y].maxdepth)
+        if (!d_info[static_cast<int>(y)].maxdepth)
             continue;
-        if (!max_dlv[y])
+        if (!max_dlv[static_cast<int>(y)])
             continue;
-        if (d_info[y].final_guardian) {
-            if (!r_info[d_info[y].final_guardian].max_num)
+        if (d_info[static_cast<int>(y)].final_guardian) {
+            if (!r_info[d_info[static_cast<int>(y)].final_guardian].max_num)
                 seiha = true;
-        } else if (max_dlv[y] == d_info[y].maxdepth)
+        } else if (max_dlv[static_cast<int>(y)] == d_info[static_cast<int>(y)].maxdepth)
             seiha = true;
 
-        fprintf(fff, _("   %c%-12s: %3d 階\n", "   %c%-16s: level %3d\n"), seiha ? '!' : ' ', d_info[y].name.c_str(), (int)max_dlv[y]);
+        fprintf(fff, _("   %c%-12s: %3d 階\n", "   %c%-16s: level %3d\n"), seiha ? '!' : ' ', d_info[static_cast<int>(y)].name.c_str(), (int)max_dlv[static_cast<int>(y)]);
     }
 }
