+++ src/io/write-diary.cpp
@@ -89,9 +89,9 @@ static QUEST_IDX write_floor(player_type *creature_ptr, concptr *note_level, cha
 	else
 	{
 #ifdef JP
-		sprintf(note_level_buf, "%d階(%s):", (int)floor_ptr->dun_level, d_info[creature_ptr->dungeon_idx].name.c_str());
+		sprintf(note_level_buf, "%d階(%s):", (int)floor_ptr->dun_level, d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str());
 #else
-		sprintf(note_level_buf, "%s L%d:", d_info[creature_ptr->dungeon_idx].name.c_str(), (int)floor_ptr->dun_level);
+		sprintf(note_level_buf, "%s L%d:", d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str(), (int)floor_ptr->dun_level);
 #endif
 		*note_level = note_level_buf;
 	}
@@ -275,16 +275,16 @@ errr exe_write_diary(player_type *creature_ptr, int type, int num, concptr note)
 	{
 		fprintf(fff, _(" %2d:%02d %20s %sの最深階%d階に到達した。\n",
 			" %2d:%02d %20s reached level %d of %s for the first time.\n"), hour, min, note_level,
-			_(d_info[creature_ptr->dungeon_idx].name.c_str(), num),
-			_(num, d_info[creature_ptr->dungeon_idx].name.c_str()));
+			_(d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str(), num),
+			_(num, d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str()));
 		break;
 	}
 	case DIARY_TRUMP:
 	{
 		fprintf(fff, _(" %2d:%02d %20s %s%sの最深階を%d階にセットした。\n",
 			" %2d:%02d %20s reset recall level of %s to %d %s.\n"), hour, min, note_level, note,
-			_(d_info[num].name.c_str(), (int)max_dlv[num]),
-			_((int)max_dlv[num], d_info[num].name.c_str()));
+			_(d_info[static_cast<int>(num)].name.c_str(), (int)max_dlv[static_cast<int>(num)]),
+			_((int)max_dlv[static_cast<int>(num)], d_info[static_cast<int>(num)].name.c_str()));
 		break;
 	}
 	case DIARY_STAIR:
@@ -302,8 +302,8 @@ errr exe_write_diary(player_type *creature_ptr, int type, int num, concptr note)
 	{
 		if (!num)
 			fprintf(fff, _(" %2d:%02d %20s 帰還を使って%sの%d階へ下りた。\n", " %2d:%02d %20s recalled to dungeon level %d of %s.\n"),
-				hour, min, note_level, _(d_info[creature_ptr->dungeon_idx].name.c_str(), (int)max_dlv[creature_ptr->dungeon_idx]),
-				_((int)max_dlv[creature_ptr->dungeon_idx], d_info[creature_ptr->dungeon_idx].name.c_str()));
+				hour, min, note_level, _(d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str(), (int)max_dlv[static_cast<int>(creature_ptr->dungeon_idx)]),
+				_((int)max_dlv[static_cast<int>(creature_ptr->dungeon_idx)], d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str()));
 		else
 			fprintf(fff, _(" %2d:%02d %20s 帰還を使って地上へと戻った。\n", " %2d:%02d %20s recalled from dungeon to surface.\n"), hour, min, note_level);
 
@@ -364,7 +364,7 @@ errr exe_write_diary(player_type *creature_ptr, int type, int num, concptr note)
 	{
         concptr to = !is_in_dungeon(creature_ptr)
 			? _("地上", "the surface")
-			: format(_("%d階(%s)", "level %d of %s"), creature_ptr->current_floor_ptr->dun_level, d_info[creature_ptr->dungeon_idx].name.c_str());
+			: format(_("%d階(%s)", "level %d of %s"), creature_ptr->current_floor_ptr->dun_level, d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str());
 		fprintf(fff, _(" %2d:%02d %20s %sへとウィザード・テレポートで移動した。\n",
 			" %2d:%02d %20s wizard-teleported to %s.\n"), hour, min, note_level, to);
 		break;
@@ -373,7 +373,7 @@ errr exe_write_diary(player_type *creature_ptr, int type, int num, concptr note)
 	{
 		concptr to = !is_in_dungeon(creature_ptr)
 			? _("地上", "the surface")
-			: format(_("%d階(%s)", "level %d of %s"), creature_ptr->current_floor_ptr->dun_level, d_info[creature_ptr->dungeon_idx].name.c_str());
+			: format(_("%d階(%s)", "level %d of %s"), creature_ptr->current_floor_ptr->dun_level, d_info[static_cast<int>(creature_ptr->dungeon_idx)].name.c_str());
 		fprintf(fff, _(" %2d:%02d %20s %sへとパターンの力で移動した。\n",
 			" %2d:%02d %20s used Pattern to teleport to %s.\n"), hour, min, note_level, to);
 		break;
