+++ src/monster-floor/monster-summon.cpp
@@ -65,7 +65,7 @@ static bool summon_specific_okay(player_type *player_ptr, MONRACE_IDX r_idx)
         && monster_has_hostile_align(player_ptr, NULL, 10, -10, r_ptr))
         return false;
 
-    if ((r_ptr->flags7 & RF7_CHAMELEON) && d_info[player_ptr->dungeon_idx].flags.has(DF::CHAMELEON))
+    if ((r_ptr->flags7 & RF7_CHAMELEON) && d_info[static_cast<int>(player_ptr->dungeon_idx)].flags.has(DF::CHAMELEON))
         return true;
 
     if (summon_specific_who > 0) {
